Refactor: Centralize and correct icon pack caching logic
This commit refactors the logic for caching icons from icon packs to fix a bug where icons were being incorrectly named and overwritten. It also streamlines the codebase by removing a redundant use case and centralizing shared logic.

### Core Changes:

-   **Correct Icon Naming**:
    -   Icons cached from icon packs are now named using the application's `componentName` (e.g., `com.package.name-com.package.name.MainActivity`) instead of just the `packageName`. This fixes a critical bug where multiple activities from the same package would overwrite each other's cached icons, leading to incorrect icons being displayed.

-   **Centralized Logic**:
    -   The logic from the now-deleted `UpdateEblanShortcutConfigsUseCase` has been extracted into a reusable `updateEblanShortcutConfigs` function in `LauncherAppsUtil.kt`. This function is now called by `SyncDataUseCase` and `ChangePackageUseCase`, reducing code duplication.
    -   A new `cacheIconPackFile` utility function has been created in `IconPackInfoUseCaseUtil.kt` to centralize the logic for finding the correct icon in an icon pack's `appfilter.xml` and saving it to the cache directory. This is used by `UpdateIconPackInfosUseCase` and `UpdateIconPackInfoByPackageNameUseCase`.

-   **Duplicate Icon Removal**:
    -   Duplicate icons provided by an icon pack are now filtered out in the `EditApplicationInfoViewModel` rather than in the `DefaultIconPackManager`. This ensures the UI layer receives a clean, distinct list of icons to present to the user.
