+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/grid/GridItem.kt
@@ -151,7 +151,10 @@ internal fun ApplicationInfoGridItem(
 
     val iconPackDirectory = File(iconPacksDirectory, iconPackInfoPackageName)
 
-    val iconPackFile = File(iconPackDirectory, data.packageName)
+    val iconPackFile = File(
+        iconPackDirectory,
+        data.componentName.replace("/", "-"),
+    )
 
     val icon = if (iconPackInfoPackageName.isNotEmpty() && iconPackFile.exists()) {
         iconPackFile.absolutePath
