+++ framework/icon-pack-manager/src/main/kotlin/com/eblan/launcher/framework/iconpackmanager/DefaultIconPackManager.kt
@@ -196,8 +196,6 @@ internal class DefaultIconPackManager @Inject constructor(
             }
         }
 
-        return iconPackInfoComponents.distinctBy { iconPackInfoComponent ->
-            iconPackInfoComponent.drawable
-        }
+        return iconPackInfoComponents
     }
 }
